[Android] MOBILE-3755 adds FF parsing and evaluation (#1307)
* stub ariship feature flags module

* fixed typo

* added feature flag info parsing

* added draft implementation
added unit test

* pr feedback
filter FF with name

* added remote data refresh

* use static device info for FF evaluation

* fixed unit tests

* fixed unit tests

* fixed ff parsing

* spotless

* updated mapping

* add device type experiments \ ff test
add evaluation options parsing \ processing
add unit tests

* static payload always returns a variant

* pr feedback

* removed unused sesion number

* spotless

---------

+++ urbanairship-core/src/main/java/com/urbanairship/audience/AudienceHash.kt
@@ -85,7 +85,7 @@ internal data class AudienceHash(
         }
     }
 
-    internal fun generate(properties: Map<String, String?>): Long? {
+    internal fun generate(properties: Map<String, String?>): ULong? {
         if (!properties.containsKey(property.jsonValue)) {
             UALog.e { "can't find device property ${property.jsonValue}" }
         }
@@ -97,7 +97,7 @@ internal data class AudienceHash(
             HashAlgorithm.FARM -> FarmHashFingerprint64::fingerprint
         }
 
-        return hashFunction.invoke("$prefix$value") % numberOfHashBuckets
+        return hashFunction.invoke("$prefix$value").toULong() % numberOfHashBuckets.toUInt()
     }
 
     override fun toJsonValue(): JsonValue {
